[STORM-175] Basic CLI
Basic CLI implementation using argparse. A client library binding with a
resource model is implemented on top of the REST API. The resource model
contains definition for triggers, rules, and actions. The CLI uses the
client library to interface with the API. Users will be able to write
shell scripts with the CLI and python scripts with the library binding
using this package.