[STORM-175] Basic CLI - Support resource update
Added support to allow update of resource types such as action and
rule. The update command requires resource ID and the modified JSON file as arguments. Since ID is provided in the argument, ID is not required in the JSON file for the update. However, if there is an ID in the JSON file and it does not match with the ID provided in the argument then an exception will be thrown.

Example:
st2 rule update 96f111045f65427cbff062f217c96f10 myrule.json