[STORM-299] Content Structure and Location
Clean up project folder structure for contents such as sensors,
triggers, rules, and actions. The following describe the changes. Repo
structure is defined and configurable via appropriate config file. Start
script for each service will create the high level paths (i.e.
/opt/stackstorm/sensors) defined in their config file as part of the startup sequence.

Repo Structure
The following is the file system structure for the repo regardless of
system vs. user generated contents.
./sensors -> sensor plugins
./sensors/webhooks -> custom webhooks and schema
./sensors/schedules -> timer schedules
./sensors/triggers -> additional triggers in JSON files to be registered
outside of sensors
./rules -> rules in JSON files to be registered
./actions -> action plugins and the JSON files required for the action
plugins to be registered

Repo Locations
Each repo location listed below is expected to have the same underlying
structure defined above in the "Repo Structure" section.
/opt/stackstorm -> Default location for user generated contents on
product installation. The content folder is removed from the path to
make this more simple (i.e. /opt/stackstorm/sensors,
/opt/stackstorm/rules, etc.).
stanley/contrib/core -> Location in the github repo for St2 contents
that will be shipped with the product.  These contents will be copied
over to /opt/stackstorm on install and then auto-registered.  We expect
users to be able to edit these contents post install.  The use of the
name "core" represent that this is a set of core contents that it's
valuable to be shipped and installed with the product.  Alternate name
is "base" for base contents or basic contents.
stanley/contrib/examples -> Location in the github repo for St2 samples
that may be of interest to the users.  These samples will be shipped
with the product but will not be auto-registered.
stanley/contrib/sandbox/packages/<SomePackageName> -> Location in the
github repo for self contained packages that we are developing (i.e.
correlation).  The <SomePackageName> is a directory with an arbitrary
name we give for the package. The structure should the same as a normal
repo structure with sensors, rules, and actions directories.  These
packages will not be shipped with the product.
stanley/contrib/sandbox/scripts -> Location in the github repo for any
scripts and tools for development purposes.