Make Ad Hoc triggers work end-to-end
I had to make a few modifications to the big picture of how RulesEngine work. I've decided to make it handle one trigger tinstance at the time (at least temporary) to avoid some complex hashing\unhashing issues in relation to trigger parameters. Personally, I think it's ok if sensor would dispatch its triggers one by one, after all we do have a queue in place for that, but if there is a valid reason to dispatch multiple events at the time, we may get back to it later.