[STORM-378] CLI for action metadata
Refactor the help message for action to display metadata. The runner
parameters is not displayed on action get so not to confuse user. The
runner parameters is displayed in action help message. Runner type get
and list are also implemented.

Examples:
st2 run local -h
st2 action execute local -h