Add sort to REST API for action execution history
The default sort order for the history records is by descending order of
the execution start_timestamp.
http://localhost:9101/history/executions

The placement of the datetimes in a range query determine the sort order automatically.
http://localhost:9101/history/executions?timestamp=20141225T000010..20141225T000019
http://localhost:9101/history/executions?timestamp=20141225T000019..20141225T000010

The sort order can be overridden by a query param named "sort".
http://localhost:9101/history/executions?sort=action,timestamp

The sort direction can be changed to descending by prepending '-'.
http://localhost:9101/history/executions?sort=action,-timestamp