Fix duplicate history record
Action execution is processed for each create and update events. Original method that process the events create history record if nonexistent. Race condition is possible where two separate events for the same action execution are processed simultaneously. The fix has the update wait for the history record and only creates a new history record on timeout.