Fix missing content pack on action update
The action content_pack is auto-assigned to default if not provided on create. This is inconsistent on update as the API throws exception on the same action metadata file. So if content_pack is not provided on update, the content_pack currently assigned to the action in the database is used.