Pin mistral client to specific version
Current trunk of mistral client is not compatible with the st2-0.5.1 branch of mistral. The change here pin the mistral client to a specific release. Move the installation of mistral before the st2 deb packages so the st2-0.5.1 branch of python-mistralclient will already satisfy the dependencies listed in the st2 requirements.txt.