Manual merge of master/14aa26b26fb8e70ef0f7a0b01d7faf6d90836b7b
New trigger should be created for each rule with different trigger parameters

Currently, we don't create a trigger for each rule. This results in parameters
being equal for all triggers. The behavior we want is to reuse the same trigger
if the parameters are same. If they are different, we want a new trigger to be
created.

* _get_trigger_api_given_rule in st2common/services/triggers now returns the right
 trigger API object.
* Add test cases to catch the bug.
* Add fixtures.

Conflicts:
	st2reactor/tests/unit/test_sensor_and_rule_registration.py