Pass runner generated auth token to Mistral
When auth is turned on, in order for Mistral to be able to launch st2 actions, the API calls to st2 from Mistral requires a token. The runner container generates a temporary token for authenticated calls. This auth token is passed to Mistral under st2_context in the environment on workbook and workflow execution.