Allow option to bypass SSL Certificate Check
The recommended behavior is to have all certificates checked for
validity before performing a clone. However, there are times when
this is outright not possible (development with a self-signed cert,
for example).

This commit updates the `packs.download` command, introducing a new
attribute `verifyssl`. By default, the behavior is to continue checking
SSL certificate validity, but now it is possible to clone with an
invalid certificate and proceed along without a care in the world.

Be careful using this... especially in the context of Infra as Code.