Repair failing pack installation
This PR fixes two bugs introduced in #1570 / #1571

* Double `.git` extension
* SSH protocol downloads via Git

In certain circumstances, the `repo_url` passed into the `packs.download`
action is returns a double `.git.git` extension. Contains fix to only
append `.git` suffix if doesn't already exist.

In #1570, the protocol for Git was changed from SSH -> HTTPS. In cases
where a git-helper is not installed (ref: st2workroom), the protocol
needs to be reverted to avoid credential prompts.