Make parent context in workflows a dict
* We need to be able to add more contextual information
about the parent to tasks in the workflow. So instead
of parent being a simple string field, I am changing
this to a dict.

* Also, liveaction_id is an internal only representation
which should not be exposed to outside. I am also
changing that behavior.

* With execution_id being in context, we need to get
liveaction corresponding to the execution. This will
subsequently vanish once all contextual info is in the dict.