Update execution object when liveaction is updated
Fixes: 1717

Several things were uncovered with this bug report.

1) A non UTF-8 string '\x82' in stdout causes mongo to throw an
exception when updating the results. We don't have a check
for validating if string is UTF-8 in document schema.
So this must be some mongo internal validation. We should
figure out if we can turn it off.

2) An exception thrown is caught by worker and liveaction
object is updated but the execution object is not. This
results in execution state shown as `running` even though
liveaction status is `failed`.

3) While fixing this, I also noticed that when an action is
canceled, we don't execute the execution object status to
`canceled` which also is consistent with the bug report.

(cherry picked from commit 8655b6bfde7509e6ff1adce3beb8735bbbcb7df1)