Fix docs mentioning 9100 and 9101 ports
There will be a lot of confusion next couple of month related to this new schema and how port and path corelate and when you should use one or the other.

The rule of thumb I used while making this commit and suggesting you should use when in doubt is rather simple:

> `https://example.org/api` is a public url and should be used for external services talking to st2 (webhooks, browsers, off-server or potentially off-serverrunning services like hubot). `http://localhost:9101/` is a private url and should be used for testing or wiring between internal components (so it is ok to use it as a mock in tests, example when testing custom auth is properly set up or to route the request from nginx to st2api).

There's a ton of potentially tricky situations where it's not immediately clear whether it should be private or public url and we'll have to deal with each of them on a case by case basis.

And yeah, if you can avoid mentioning st2api url in examples that are not directly tied to wiring up the system, use some other url instead. 'http://httpbin.org/get' is a nice general purpose example url.