Add the `extras` field to action aliases
Different chat adapters have special capabilities which are sometimes
very handy and nice. Slack has attachment colors and parsing/unfurling
options for better formatting; HipChat has Connect features; and so on.
It’s hard for users to add support for those right now, because you
have to modify the schema and the DB connector in st2, and only then
submit a PR to `hubot-stackstorm` utilizing the new parameter. More
importantly, if we had a lot of requests for this, our schema would be
overblown with rarely used parameters.

A solution for this,—which has been discussed some weeks ago, but I
only got around to it now,—is an `extras` object inside the alias with
a “free” schema. It will allow people to easily add support for extra
features without modifying `st2` (which would imply only getting a
feature out in the next `st2` release and complicating the schema) and
with relative ease. While this is not ideal architecturally (a reason
the schema exists in the first place is to have order, and this breaks
it a bit), I feel like pros outweigh the cons, and it’s worth to
introduce this. We still maintain strict schema outside `extras`.

Another neat feature this PR includes is rendering extras with Jinja
sharing the context with `ack` or `result`, depending on the placement.
E.g. different colors depending on the execution status.

This can also be useful for possible usage of action aliases outside
chat (e-mails, sms, tweets, and whatnot).

An example:
```
name: "tesla.honk_horn"
action_ref: "tesla.honk_horn"
description: "Honk your horn"
formats:
- “honk-honk”
result:
format: "**Honk Honk**"
extras:
color: “<% if result.succeeded %> green <% else %> red <% endif %>”
```