Add support for Jinja template notation for default parameter values for parameters which type is object.
Note 1: We already supported it for strings and list, but we didn't support it for
objects.

Note 2: This change only adds support for objects with single level of nesting.
This is done intentionally to avoid potentially infinite loops and resource
exhaustion. It's the same story with lists - we only support rendering direct
list items.