Added handler for single-element arrays without brackets
Issue #3157 (https://github.com/StackStorm/st2/issues/3157) reported
a problem with using single-element arrays as an action parameter thusly:

	st2 run foopack.fooaction arrayparam=51

This was because the transform_array function in actions.py first tries to
load the given value as JSON. In the example above, 51 is a valid int in JSON,
so this results in an error because of a mismatch between what the API expected
(array) and what was actually sent (int).

Note that this isn't a problem if the single element in this array is a string;
json.loads handles that just fine. However, if it's an int, this bug appears.

To get around this, the user **could** run something like:

	st2 run foopack.fooaction arrayparam=[51]

However, this is a bit unintuitive. Because of the action metadata, we already know
that this **should** be a list, so we can force it to be. So, this commit adds a
step to check if the value is an int, and if so, force it to be a Python
list. This allows the user to also use the more intutive former example.

Signed-off-by: Matt Oswalt <oswaltm@brocade.com>