Support additional path requirements
The additional check helps avoid the situation when static portion of the path gets resolved as dynamic due to lack of ordering or priority during patch resolving.

An example of that situation will be an attempt to resolve a path (`/rules/views`) against two matching endpoints (`/rules/views` and `/rules/{ref_or_id}`). Depending on the dictionary order, the path may be resolved to either of two. The additional requirement helps us to stear the algorithm to a particular resolution.