Total re-vamp of login and whoami tests
The previous version of these tests were a bit too reliant on mock,
so this approach instead works with as few mocks as possible.

Naturally, we have to keep the mock for the st2 API interactions, but
these tests run the `st2 login` and `st2 whoami` commands using
real config files, as opposed to a mocked ConfigParser.

Signed-off-by: Matt Oswalt <oswaltm@brocade.com>