Add support for nullable types
The flavor of jsonschema used in openapi prohibits use of `type: null`. At least one of our endpoints may get null value as an input. At that moment I'm not comfortable to change existing API behaviour so I had to add `x-nullable` schema properly that skips type validation if it's set to True and value of the prarmeter is None.