Add missing openapi definition for API key model.
We didn't have one so the new RBAC tests failed. This undercovered a
bigger issue with lack of good tests - we were missing a definition
so the whole endpoint worked incorrectly, but because of lack of
static typing and over-use of getattr() existing tests passed even
though if functionality was half broken.