Modify CORS behaviour to work around browser security limitations
Generally, we should not allow using cookes for authentication without expressingly setting a limited number of allowed origins.

However, due to inconsistencies in browser implementations and the responsiveness of our community to the whole concept of CORS over the last few years, I'm inclined to say "the hell with it".

This is a little hack that allows us to work around the problem described in https://stackoverflow.com/questions/19743396/cors-cannot-use-wildcard-in-access-control-allow-origin-when-credentials-flag-i.