Fix cherry pick conflicts.
Fix invocation of Mistral workflow from Action Chain

If the parent action chain contains jinja expression in the execution parameters, the jinja expression will be included in the context begin passed into the Mistral subworkflow. Mistral will try to evaluate the expression when it is evaulating input args and if the expression contains references to local data context, the evaulation fails because the data context is already out of scope. This patch converts the jinja expression into raw jinja blocks when passing thru the context.