Add ability to pre-declare all the used queues on service set up.
This is required when using worker model where all the services which
consume from queues might not already be online before the producers.

Because of the implementation details, this is not required with
RabbitMQ transport, but it's requires with some other transports such
as Redis one.

This feature can be enabled by setting messaging.predeclare_queues
st2.conf config option to True (False by default).