Implement pause and resume for Action Chain
Action chain can be paused and resumed. Pause and resume will cascade down to subworkflows. If pause is initiated from a subworkflow, the pause will cascade up to the parent action chain. On the contrary, if resume is initiated from a subworkflow, the resume will not cascade up to the parent action chain.