Fix a bug (race) related to reading from process stdout and stderr when read_{stdout,stderr}_handler_func functionality is used.
We didn't correctly wait on green threads to finish reading from stdout,
stderr stream so in case of a very short running action, we would
sometimes miss some of the output of the action.

Also make sure we kill green threads which read from those streams when
action timeout is reached.