Update execution output stream to utilize a single "generic" collection for storing the output instead of utilizing two collections.
This approach makes whole thing more generic and makes it more
future-proof and work with other runner which doesn't necessary have two
streams (stdout, stderr).

Now "type" argument on the document stores type of the output and it
means each runner execution can produce as little or as many output
types as it needs.

Most of the existing runners will produce two type of output (stdout,
stderr), but it's likely that in the future we will have other runners
which will just produce one type of the output.