Make all queues used by stream service auto_delete=True.
This way they get automatically deleted when stream is shut down. We use
random names for those queues so if we don't do that, they will stay
around when stream is restart (new queues get created on restart) and
nothing will consume from them hence we are left with zombie queues.