Fix parameters in chain context of Mistral workflow
This patch is a follow up to PR https://github.com/StackStorm/st2/pull/3609. Jinja expression in the parameters of tasks in Action Chain is being evaluated in Mistral. The attribute for the Action Chain task parameters can be named "params" and "parameters". The previous PR fixed the case where the task parameters is named "params". This commit fixes the case where it's named "parameters".