Fixed bug in ActionChain runner where context_result wasn't updated properly
Inquiries play a bit of surgery with the liveactions behind the scenes, so when an ActionChain
resumes from a Inquiry pause, it correctly retrieves an updated result when it detects a state
change it didn't previously know about.

However, it doesn't update context_result, which is what ultimately gets passed into the Jinja
renderer. So this makes sure the appropriate task results is updated in context_result

Signed-off-by: Matt Oswalt <matt@keepingitclassless.net>