Update Python runner so it's now standalone.
This means that config is now passed to the runner and runner doesn't
need access to database anymore.

Config abstraction has now been moved one level higher and it's loaded
inside the runner container service. This abstraction is better anyway
because we can eventually also easily add support for configs to other
runners.