Fix edge case for workflows stuck in running state
There is an edge case where mistral receives connection error from the st2 API when requesting action execution. When this happens, the action execution is requested twice. The first action execution is requested, the record is written into the MongoDB, but it is not published to the action runners. The second action execution is scheduled and the mistral workflow reach completion. The mistral querier for the st2resultstracker will see the first action execution in a non-completed state and assume the workflow is still running. This patch identifies this specific edge case and ignores the duplicate action execution.