Add MY_EXECUTION_VIEW Permission Type
MY_EXECUTION_VIEW is a limited form of EXECUTION_VIEW.
It allows a user to only view executions that they started.

This is important because, though ACTION_EXECUTE implies ACTION_VIEW,
it does not imply EXECUTION_VIEW. If a user had a role that gave them
ACTION_VIEW on a particular action, then they would get the implicit
EXECUTION_VIEW permission on all executions for that action even if they
did not start the action. The implicit permission from ACTION_EXECUTE
to ACTION_VIEW and from ACTION_VIEW to EXECUTION_VIEW make sense, but
adding an additional implicit permission from ACTION_EXECUTE to
EXECUTION_VIEW would, perhaps, open up more data than intended. This
resolves that by introducing a concept of execution ownership.

MY_EXECUTION_VIEW can be a global permission, or it can apply to a pack
or a particular action in a pack. As a global permission, a user can
view any execution that they started. As a pack-level permission, a user
can view any action executions when the associated action is part of a
particular pack and the execution was started by that same user. As an
action-level permission, the user is allowed to see the executions for a
particular action if they started that execution.

Example roles:
---
name: "view_my_executions"
description: "Allow a user to view any of their own executions"
permission_grants:
  -
    permission_types:
      - "my_execution_view"
---
name: "my_linux_actions"
description: "Allow a user to execute linux actions and view only the linux actions they started"
permission_grants:
  -
    resource_uid: "pack:linux"
    permission_types:
      - "action_execute"
      - "my_execution_view"
---
name: "my_check_loadavg"
description: "Allow a user to execute check_loadavg actions and view only the executions they started"
permission_grants:
  -
    resource_uid: "action:linux:check_loadavg"
    permission_types:
      - "action_execute"
      - "my_execution_view"

NOTE: This only adds the MY_EXECUTION_VIEW permission. It does not make
it an implicit permission of any other permissions. It might make sense
to make ACTION_EXECUTE imply MY_EXECUTION_VIEW, but this commit does not
do that, leaving it as an opportunity for future development.