Fix support for mistral reverse workflows
Reverse workflows require the task_name parameter[1] to indicate which
task is the target task for the workflow (ie what the last task should
be on successful completion of the workflow, all other tasks being
determined by the requirements of the various tasks).

This adds support for the task_name parameter to the mistral runner and
fixes the parameter name in the runner config.

[1] https://docs.openstack.org/mistral/latest/api/v2.html#mistral.api.controllers.v2.resources.Execution.params

Fixes #3994