Refactor the callback module for post run to be more generic
Refactor the callback module for runner post run to be more generic. Currently, the callback takes a REST endpoint as input argument. The callback module is abstracted to take other input arguments. In this patch, the entire liveaction database object is passed to avoid overcrowding the method signature. Unfortunately, the liveaction object that the runner reference to is out dated by the time the runner makes the post run and there are more than one references of the object in the runner container. The runner container and runner is refactored to ensure references objects are used and updated.