Add use_object_ids argument to the fixtures loader.
When this argument is True, we use object ids (primary keys) from
fixture files when saving objects in the database.

This allows us to re-create reproducible tests in cases when tests rely
on fixed ids without needing to mock a lot of things (mocking is error
prone and not robust in a lot of scenarios).