Add single sensor mode (sensorcontainer.single_sensor_mode config option) to the sensor container.
When running in this mode, it's assumed that sensor process life cycle
and failover is handled by an external service so the sensor process
container exits immediately when a server crashes.

This is useful in scenarios where sensor process life cycle and failover
is handled by external service such as kubernetes.