Fix race between task and action execution
Move the update of the task execution before the action execution is requested. Currently, the task execution is updated after the request and there are cases where the action execution completed very quickly and the completion of the action execution is processed first. When completion is processed, the task execution is also updated and this causees conflict with the previous task execution update.