Add a workaround for eventlet WSGI http server.
eventlet WSGI HTTP server sets Content-Type to text/plain if one is not
explicitly provided in the request and this brakes API endpoint which
accept application/json and where body is optional.

NOTE: This only affects eventlet wsgi server and work around is only
applied when not running under gunicorn (aka when running in standalone
mode).