Fix make tests in MakeFile
The st2client.egg-info is included in the list of components and that caused errors when making requirements and running pylints on the components. The make tests command would abort at the pylint step and stop running unit tests. The wildcard function to identify the list of st2 components is replaced with a shell command to run ls and grep the entries that start with st2 but not end with .egg-info.