Change a default value of system.validate_trigegr_parameters and system.validate_trigger_payload config option to True.
This means we now perform trigger payload validation when trigger is
dispatched via webhook / sensor and when a trigger is created via rule.

This can be disabled via those two config options, but it's recommended
to be turned on since it can catch various typos and issues which would
otherwise be silently ignored.