Update "setup_virtualenv" function so it also sets / changes owner group of the pack virtualenv directory to the value defined in the config (content.pack_group, defaults to st2packs).
This way the end result is consistent also when running
st2-pack-setup-virtualenv and st2ctl reload --register-setup-virtualenvs
command as a user which is not part of st2packs group.

Keep in mind that this step requires sudo access, but if sudo is not
available, or group ownership change step fails, the command itself
doesn't fail.

This means that sudo access for st2-pack-{download,setup-virtualenv,install}
commands is optional.