Add new ?output_type=[raw|event_source] and ?existing_only=[true|false] query parameter filters to the action execution output API endpoint.
With ?output_type user can control if they want to receive the "raw"
output (default) or "event_source" output. Keep in mind that
"event_source" format might result in a slow response if action has
already produced a lot of data, because it means we need to call
json_encode on each ActionOutpuDB object.

With ?existing_only user can control if they want the API endpoint to
only return existing output produced by the execution so far and return
/ close the connection immediately. Keep in mind that this only applies
to the executions which are still running.