Update the code to perform include and exclude attributes filtering in the router.
This way the filtering is performed at the very end, before returning
raw response to the user.

Filtering on the router makes more sense compared to filtering on the
API model because filtering the object might make it an invalid API
model (aka object won't pass API model validation anymore) and filtering
is purerly a property of the response and not the API model.