Fix race between processing a pending execution and pausing parent
If the action execution goes into pending state at the start of the inquiry, then its parent workflow will also be paused. Previously, the pause request is executed in the run method before the pending status is updated for the action execution. This lead to race where the parent is stuck in pausing because the action execution is still running. The pause request is moved to the post run method in this patch so that the pending status for the action execution is already updated after run but before post run.