Fix inquiry regression for mistral
Move the post run of base class until after the pause request in the inquiry runner post run method. The base post run makes a callback to Mistral. We want to complete any work in st2 before the mistral callback. Add the pending status to the inactive check in action service.