Add CLI for workflow inspection
Add `st2 workflow inspect` command in st2client to inspect workflow. The command takes either a file path or action ref as input. The use of action ref is only supported if the action is locally installed under /opt/stackstorm/packs. The command then read the content from the file or entry point of the action and pass the workflow definition as text to the API for inspection. Errors found during inspection will be printed to stdout.