Optimize POST /v1/executions API endpoint by removing a number of duplicated queries.
Instead of retrieving the same items over and over again, we re-use
existing objects which are retrieved beforehand.

This goes from 14 queries during that API endpoint operation to 3
queries.

All of those queries are relatively simply get object based on the value
of index field, yet they are still wasteful. We are trading off a bit of
code readability for performance.