Fix an issue with timers engine config options not working as advertised.
We advertise "timersengine" section as new and preferred way of
configuring timers engine service and "timer" config options are
deprecated.

Because of the way how the code was structured[1], even if user would
specify only "timersengine" section in st2.conf, those options would
not be used, but default options from config.py for "timer" section
would be used.

This pull request fixes that to work as advertised / inteded - "timer"
section should only be used if explicitly specified in st2.conf and
timersengine should have precedence over it.

[1]: https://github.com/StackStorm/st2/blob/0245a18198babafe2ab0abde31791cf526435234/st2reactor/st2reactor/cmd/timersengine.py#L59