Fix race and random failures with unit tests for concurrency policy
Fix race and random failures with unit tests for concurrency policy by using the scheduling only mock publisher and handling the processing of the scheduling queue manually. Use the noop coordinator driver to prevent blocking in the process. The lock mechanism should be unit tested elsewhere. The separation of concerns should improve the robustness of these tests.