Reset the global coordinator on class setup and tear down
In the unit tests for the concurrency policy, reset the global coordinator in setup and tear down. Since we are using the noop coordinator in these tests, if they are not reset on setup, the global coordinator will return the default with zake driver. The reset at tear down is to ensure other tests are not affected.