Fix task with items but items list is empty for orquesta workflows
If the items list is empty, on transition to the task with items, orquesta fails to render the action specs correctly. For this use case, the orquesta conductor will now return the task spec with the list of action specs empty. The st2 workflow engine will recognize and handle this appropriately to set the task execution to completion in the conductor as well as the database records to progress the workflow execution to the next task(s).