Rename python_version attribute on PackDB model to python_versions and make it a list.
This attribute specifies which major Python versions the pack works with
(2, 3).

If this attribute is not specified in pack.yaml, we perform no Python
version compatibility check when installing and registering the pack.

The check can also be skipped by using --force install flag, but that
is at user's own risk and packs are not guaranteed to work.

Update affected tests.