Update service_setup service entry / bootstrap code so we exclude log messages with log level "AUDIT" if log level is set to INFO or higher.
This way we avoid issues with duplicate AUDIT messages in production
deployments.

In production deployments default log level is set to INFO which means
that all service log files will also contain AUDIT log messages because
AUDIT level if the highest.

This is not desired, because we already log AUDIT log messages in a
dedicated AUDIT log file.

NOTE: Audit messages will still go in the service log file is log level
is set to DEBUG (that's desired during debugging).