Treat any statsd metric driver related errors as non-fatal.
Underneath, statsd uses UDP protocol which is "fire and forget" which
means all the errors are non-fatal and don't propagate back to the
application. There is an edge case with how statsd library establishes
connection though. If hostname is used instead of an IP address and this
hostname can't be resolved, this DNS resolution error will propagate all
the way up and break the application.

This behavior is of course not desired (all metric related operation
errors should be non-fatal) and this PR fixes that. We simply ignore
those errors and log them.