Remove policy-delayed status to avoid bouncing between delayed statuses
Currently, when concurrency policy decide the action execution should be delayed, the action execution status is changed to policy-delayed. The scheduler moves the action execution to delayed status. If there is a lot of traffic and the action execution is continued to be delayed, the bouncing between policy-delayed and delayed causes the log to be inflated with records of the status changes. This patch remove the policy-delayed status altogether and just use the delayed status so there's no constant and unncessary status changes.