Disallow decrypt_kv jinja filter for fields marked as secret
This commit addresses the issue where in pack config if fields are marked as
secret: true and if user specifies jinja expression with filter decrypt_kv,
the values are decrypted twice. This is due to the fact that for all fields
marked as secret, the values are auto decrypted. Specifying an additional
decrypt_kv filter causes issue.

The commit raises exceptions if decrypt_kv is specified for any fields marked secret.