Add support for pluggable concurrency for the util.green.shell module.
This way we can use either eventlet or gevent, depending on the
application (e.g. gRPC doesn't support eventlet so we need to use gevent).

Also fix a bug with where we set returncode on the process. If we set it
before calling kill() it will get overwritten inside kill() method.