Fix logic in identifying orphaned workflow executions
In the query to identify task executions, add id__ne filter so that the action execution for the workflow execution is not included.  This fixes the case where there is no task executions for the workflow execution. If there are task executions for the workflow execution, fix the if statement that identifies if there is any active tasks and if there are not active tasks whether the most recently completed task has passed expiry datetime.