Handle unknown except during cancelation
Instead of rethrowing the exception, capture and log the error and traceback, and return the error in the result. Then let the cancelation continue to process children and determine final status. If the exception is rethrown, the workflow execution will be stuck in a canceling status. Add unit test to cover this scenario.