Get rid if automatic auto conversion / de-referencing of embedded fields inside the base mongoengine document class.
This process is very slow and adds a tons of overhead since it calls
"to_python()" on every model field even if the field doesn't need to be
referenced.

With this approach we manually de-reference only the fields which need
to be dereferenced / converted.

The most ideal and faster approach would be to work directly with
dictionaries as returned by pymongo (this way we could avoid the whole
overhead), but this is a compromise.