Ensure that py3 tests use python3.6 virtualenv
There were several issues with the python3.6 virtualenvs.
- The Makefile defaults PYTHON_VERSION to python2.7 causing some
  virtualenvs to be built with 2.7 instead of 3.6.
- Once a virtualenv was built with 2.7 it was cached for later builds.
- The ci-py3-unit task was running tox for both unit and packs tests.
  Hiding discrete steps in the same task obscured the lack of
  pre-building one of the tox virtualenvs. It also made it hard to
  understand how the job name matched with the TASK/make target.

So, this updates the travis jobs to add PYTHON_VERSION and update the
job TASK to separate ci-py3-unit and ci-py3-packs-tests make targets.
Several other places that use TASK var were adjusted to clarify WHY they
were selecting different CI TASKs, thus allowing TASK to be adjusted or
reorganized without editing so many different files.

In one case, an if statement was selecting all of the TASK options, so
the condition was removed to leave just the command invocation.

TASK & Makefile to separate py3-unit and
py3-packs-tests. And updates several places that use the TASK var to
make it clear WHY they are selecting particular TASKs.